Module google_logging2_beta1::api

Structs

BucketOptions
BucketOptions describes the bucket boundaries used to create a histogram for the distribution. The buckets can be in a linear sequence, an exponential sequence, or each bucket can be specified explicitly. BucketOptions does not include the number of values in each bucket.A bucket has an inclusive lower bound and exclusive upper bound for the values that are counted for that bucket. The upper bound of a bucket must be strictly greater than the lower bound. The sequence of N buckets for a distribution consists of an underflow bucket (number 0), zero or more finite buckets (number 1 through N - 2) and an overflow bucket (number N - 1). The buckets are contiguous: the lower bound of bucket i (i > 0) is the same as the upper bound of bucket i - 1. The buckets span the whole range of finite values: lower bound of the underflow bucket is -infinity and the upper bound of the overflow bucket is +infinity. The finite buckets are so-called because both bounds are finite.

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

EntryListCall
Lists log entries. Use this method to retrieve log entries that originated from a project/folder/organization/billing account. For ways to export log entries, see Exporting Logs.

EntryMethods
A builder providing access to all methods supported on entry resources. It is not used directly, but through the Logging hub.

EntryWriteCall
Writes log entries to Logging. This API method is the only way to send log entries to Logging. This method is used, directly or indirectly, by the Logging agent (fluentd) and all logging libraries configured to use Logging. A single request may contain log entries for a maximum of 1000 different resources (projects, organizations, billing accounts or folders)

Explicit
Specifies a set of buckets with arbitrary widths.There are size(bounds) + 1 (= N) buckets. Bucket i has the following boundaries:Upper bound (0 <= i < N-1): boundsi Lower bound (1 <= i < N); boundsi - 1The bounds field must contain at least one element. If bounds has only one element, then there are no finite buckets, and that single element is the common boundary of the overflow and underflow buckets.

Exponential
Specifies an exponential sequence of buckets that have a width that is proportional to the value of the lower bound. Each bucket represents a constant relative uncertainty on a specific value in the bucket.There are num_finite_buckets + 2 (= N) buckets. Bucket i has the following boundaries:Upper bound (0 <= i < N-1): scale * (growth_factor ^ i). Lower bound (1 <= i < N): scale * (growth_factor ^ (i - 1)).

HttpRequest
A common proto for logging HTTP requests. Only contains semantics defined by the HTTP specification. Product-specific logging information MUST be defined in a separate message.

LabelDescriptor
A description of a label.

Linear
Specifies a linear sequence of buckets that all have the same width (except overflow and underflow). Each bucket represents a constant absolute uncertainty on the specific value in the bucket.There are num_finite_buckets + 2 (= N) buckets. Bucket i has the following boundaries:Upper bound (0 <= i < N-1): offset + (width * i). Lower bound (1 <= i < N): offset + (width * (i - 1)).

ListLogEntriesRequest
The parameters to ListLogEntries.

ListLogEntriesResponse
Result returned from ListLogEntries.

ListLogMetricsResponse
Result returned from ListLogMetrics.

ListMonitoredResourceDescriptorsResponse
Result returned from ListMonitoredResourceDescriptors.

ListSinksResponse
Result returned from ListSinks.

LogEntry
An individual entry in a log.

LogEntryOperation
Additional information about a potentially long-running operation with which a log entry is associated.

LogEntrySourceLocation
Additional information about the source code location that produced the log entry.

LogMetric
Describes a logs-based metric. The value of the metric is the number of log entries that match a logs filter in a given time interval.Logs-based metric can also be used to extract values from logs and create a a distribution of the values. The distribution records the statistics of the extracted values along with an optional histogram of the values as specified by the bucket options.

LogSink
Describes a sink used to export log entries to one of the following destinations in any project: a Cloud Storage bucket, a BigQuery dataset, or a Cloud Pub/Sub topic. A logs filter controls which log entries are exported. The sink must be created within a project, organization, billing account, or folder.

Logging
Central instance to access all Logging related resource activities

MetricDescriptor
Defines a metric type and its schema. Once a metric descriptor is created, deleting or altering it stops data collection and makes the metric type's existing data unusable.

MetricDescriptorMetadata
Additional annotations that can be used to guide the usage of a metric.

MonitoredResource
An object representing a resource that can be used for monitoring, logging, billing, or other purposes. Examples include virtual machine instances, databases, and storage devices such as disks. The type field identifies a MonitoredResourceDescriptor object that describes the resource's schema. Information in the labels field identifies the actual resource and its attributes according to the schema. For example, a particular Compute Engine VM instance could be represented by the following object, because the MonitoredResourceDescriptor for "gce_instance" has labels "instance_id" and "zone": { "type": "gce_instance", "labels": { "instance_id": "12345678901234", "zone": "us-central1-a" }}

MonitoredResourceDescriptor
An object that describes the schema of a MonitoredResource object using a type name and a set of labels. For example, the monitored resource descriptor for Google Compute Engine VM instances has a type of "gce_instance" and specifies the use of the labels "instance_id" and "zone" to identify particular VM instances.Different APIs can support different monitored resource types. APIs generally provide a list method that returns the monitored resource descriptors used by the API.

MonitoredResourceDescriptorListCall
Lists the descriptors for monitored resource types used by Logging.

MonitoredResourceDescriptorMethods
A builder providing access to all methods supported on monitoredResourceDescriptor resources. It is not used directly, but through the Logging hub.

MonitoredResourceMetadata
Auxiliary metadata for a MonitoredResource object. MonitoredResource objects contain the minimum set of information to uniquely identify a monitored resource instance. There is some other useful auxiliary metadata. Monitoring and Logging use an ingestion pipeline to extract metadata for cloud resources of all types, and store the metadata in this message.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the Logging hub.

ProjectMetricCreateCall
Creates a logs-based metric.

ProjectMetricDeleteCall
Deletes a logs-based metric.

ProjectMetricGetCall
Gets a logs-based metric.

ProjectMetricListCall
Lists logs-based metrics.

ProjectMetricUpdateCall
Creates or updates a logs-based metric.

ProjectSinkCreateCall
Creates a sink that exports specified log entries to a destination. The export of newly-ingested log entries begins immediately, unless the sink's writer_identity is not permitted to write to the destination. A sink can export log entries only from the resource owning the sink.

ProjectSinkDeleteCall
Deletes a sink. If the sink has a unique writer_identity, then that service account is also deleted.

ProjectSinkGetCall
Gets a sink.

ProjectSinkListCall
Lists sinks.

ProjectSinkUpdateCall
Updates a sink. This method replaces the following fields in the existing sink with values from the new sink: destination, and filter.The updated sink might also have a new writer_identity; see the unique_writer_identity field.

WriteLogEntriesRequest
The parameters to WriteLogEntries.

WriteLogEntriesResponse
Result returned from WriteLogEntries. empty

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.