Struct google_logging2_beta1::EntryMethods

pub struct EntryMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on entry resources. It is not used directly, but through the Logging hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_logging2_beta1 as logging2_beta1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use logging2_beta1::Logging;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Logging::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `list(...)` and `write(...)`
// to build up your call.
let rb = hub.entries();

Methods

impl<'a, C, A> EntryMethods<'a, C, A>

fn list(&self, request: ListLogEntriesRequest) -> EntryListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists log entries. Use this method to retrieve log entries from Cloud Logging. For ways to export log entries, see Exporting Logs.

Arguments

request - No description provided.

fn write(&self, request: WriteLogEntriesRequest) -> EntryWriteCall<'a, C, A>

Create a builder to help you perform the following task:

Writes log entries to Cloud Logging. All log entries in Cloud Logging are written by this method.

Arguments

request - No description provided.