Struct google_logging2_beta1::ListLogEntriesRequest

pub struct ListLogEntriesRequest {
    pub filter: Option<String>,
    pub project_ids: Option<Vec<String>>,
    pub order_by: Option<String>,
    pub page_size: Option<i32>,
    pub page_token: Option<String>,
}

The parameters to ListLogEntries.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list entries (request)

Fields

filter

Optional. An advanced logs filter. The filter is compared against all log entries in the projects specified by projectIds. Only entries that match the filter are retrieved. An empty filter matches all log entries.

project_ids

Required. One or more project IDs or project numbers from which to retrieve log entries. Examples of a project ID: "my-project-1A", "1234567890".

order_by

Optional. How the results should be sorted. Presently, the only permitted values are "timestamp" (default) and "timestamp desc". The first option returns entries in order of increasing values of LogEntry.timestamp (oldest first), and the second option returns entries in order of decreasing timestamps (newest first). Entries with equal timestamps are returned in order of LogEntry.insertId.

page_size

Optional. The maximum number of results to return from this request. Fewer results might be returned. You must check for the nextPageToken result to determine if additional results are available, which you can retrieve by passing the nextPageToken value in the pageToken parameter to the next request.

page_token

Optional. If the pageToken request parameter is supplied, then the next page of results in the set are retrieved. The pageToken parameter must be set with the value of the nextPageToken result parameter from the previous request. The values of projectIds, filter, and orderBy must be the same as in the previous request.