Struct google_logging2_beta1::ListLogEntriesResponse

pub struct ListLogEntriesResponse {
    pub next_page_token: Option<String>,
    pub project_id_errors: Option<HashMap<String, Status>>,
    pub entries: Option<Vec<LogEntry>>,
}

Result returned from ListLogEntries.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list entries (response)

Fields

next_page_token

If there are more results than were returned, then nextPageToken is included in the response. To get the next set of results, call this method again using the value of nextPageToken as pageToken.

project_id_errors

If partial_success is true, contains the project ids that had errors and the associated errors.

entries

A list of log entries.