Struct google_logging2_beta1::LogEntry

pub struct LogEntry {
    pub http_request: Option<HttpRequest>,
    pub resource: Option<MonitoredResource>,
    pub severity: Option<String>,
    pub text_payload: Option<String>,
    pub timestamp: Option<String>,
    pub labels: Option<HashMap<String, String>>,
    pub proto_payload: Option<HashMap<String, String>>,
    pub json_payload: Option<HashMap<String, String>>,
    pub log_name: Option<String>,
    pub insert_id: Option<String>,
    pub operation: Option<LogEntryOperation>,
}

An individual entry in a log.

This type is not used in any activity, and only used as part of another schema.

Fields
http_request: Option<HttpRequest>

Optional. Information about the HTTP request associated with this log entry, if applicable.

resource: Option<MonitoredResource>

Required. The monitored resource associated with this log entry. Example: a log entry that reports a database error would be associated with the monitored resource designating the particular database that reported the error.

severity: Option<String>

Optional. The severity of the log entry. The default value is LogSeverity.DEFAULT.

text_payload: Option<String>

The log entry payload, represented as a Unicode string (UTF-8).

timestamp: Option<String>

Optional. The time the event described by the log entry occurred. If omitted, Cloud Logging will use the time the log entry is written.

labels: Option<HashMap<String, String>>

Optional. A set of user-defined (key, value) data that provides additional information about the log entry.

proto_payload: Option<HashMap<String, String>>

The log entry payload, represented as a protocol buffer. You can only use protoPayload values that belong to a set of approved types.

json_payload: Option<HashMap<String, String>>

The log entry payload, represented as a structure that is expressed as a JSON object.

log_name: Option<String>

Required. The resource name of the log to which this log entry belongs. The format of the name is "projects/ /logs/". Examples: "projects/my-projectid/logs/syslog", "projects/1234567890/logs/library.googleapis.com%2Fbook_log". The log ID part of resource name must be less than 512 characters long and can only include the following characters: upper and lower case alphanumeric characters: [A-Za-z0-9]; and punctuation characters: forward-slash, underscore, hyphen, and period. Forward-slash (/) characters in the log ID must be URL-encoded.

insert_id: Option<String>

Optional. A unique ID for the log entry. If you provide this field, the logging service considers other log entries in the same log with the same ID as duplicates which can be removed. If omitted, Cloud Logging will generate a unique ID for this log entry.

operation: Option<LogEntryOperation>

Optional. Information about an operation associated with the log entry, if applicable.