Struct google_logging2_beta1::LogMetric

pub struct LogMetric {
    pub filter: Option<String>,
    pub name: Option<String>,
    pub description: Option<String>,
}

Describes a logs-based metric. The value of the metric is the number of log entries that match a logs filter.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

metrics create projects (request|response)

metrics get projects (response)

metrics update projects (request|response)

Fields

filter

An advanced logs filter. Example: "logName:syslog AND severity>=ERROR".

name

Required. The client-assigned metric identifier. Example: "severe_errors". Metric identifiers are limited to 1000 characters and can include only the following characters: A-Z, a-z, 0-9, and the special characters _-.,+!*',()%/\. The forward-slash character (/) denotes a hierarchy of name pieces, and it cannot be the first character of the name.

description

A description of this metric, which is used in documentation.