Struct google_logging2_beta1::LogSink

pub struct LogSink {
    pub filter: Option<String>,
    pub destination: Option<String>,
    pub name: Option<String>,
    pub output_version_format: Option<String>,
}

Describes a sink used to export log entries outside Cloud Logging.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

sinks get projects (response)

sinks update projects (request|response)

sinks create projects (request|response)

Fields

filter

An advanced logs filter. Only log entries matching that filter are exported. The filter must be consistent with the log entry format specified by the outputVersionFormat parameter, regardless of the format of the log entry that was originally written to Cloud Logging. Example (V2 format): "logName=projects/my-projectid/logs/syslog AND severity>=ERROR".

destination

The export destination. See Exporting Logs With Sinks. Examples: "storage.googleapis.com/a-bucket", "bigquery.googleapis.com/projects/a-project-id/datasets/a-dataset".

name

Required. The client-assigned sink identifier. Example: "my-severe-errors-to-pubsub". Sink identifiers are limited to 1000 characters and can include only the following characters: A-Z, a-z, 0-9, and the special characters _-..

output_version_format

The log entry version to use for this sink's exported log entries. This version does not have to correspond to the version of the log entry when it was written to Cloud Logging.