Struct google_logging2_beta1::ProjectMethods

pub struct ProjectMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on project resources. It is not used directly, but through the Logging hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_logging2_beta1 as logging2_beta1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use logging2_beta1::Logging;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Logging::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `logs_delete(...)`, `metrics_create(...)`, `metrics_delete(...)`, `metrics_get(...)`, `metrics_list(...)`, `metrics_update(...)`, `sinks_create(...)`, `sinks_delete(...)`, `sinks_get(...)`, `sinks_list(...)` and `sinks_update(...)`
// to build up your call.
let rb = hub.projects();

Methods

impl<'a, C, A> ProjectMethods<'a, C, A>

fn sinks_list(&self, project_name: &str) -> ProjectSinkListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists sinks.

Arguments

projectName - Required. The resource name of the project containing the sinks. Example: "projects/my-logging-project".

fn logs_delete(&self, log_name: &str) -> ProjectLogDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a log and all its log entries. The log will reappear if it receives new entries.

Arguments

logName - Required. The resource name of the log to delete. Example: "projects/my-project/logs/syslog".

fn sinks_get(&self, sink_name: &str) -> ProjectSinkGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a sink.

Arguments

sinkName - The resource name of the sink to return. Example: "projects/my-project-id/sinks/my-sink-id".

fn metrics_get(&self, metric_name: &str) -> ProjectMetricGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a logs-based metric.

Arguments

metricName - The resource name of the desired metric. Example: "projects/my-project-id/metrics/my-metric-id".

fn sinks_update(&self, request: LogSink, sink_name: &str) -> ProjectSinkUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates or updates a sink.

Arguments

request - No description provided.

sinkName - The resource name of the sink to update. Example: "projects/my-project-id/sinks/my-sink-id". The updated sink must be provided in the request and have the same name that is specified in sinkName. If the sink does not exist, it is created.

fn metrics_delete(&self, metric_name: &str) -> ProjectMetricDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a logs-based metric.

Arguments

metricName - The resource name of the metric to delete. Example: "projects/my-project-id/metrics/my-metric-id".

fn sinks_create(&self, request: LogSink, project_name: &str) -> ProjectSinkCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a sink.

Arguments

request - No description provided.

projectName - The resource name of the project in which to create the sink. Example: "projects/my-project-id". The new sink must be provided in the request.

fn metrics_create(&self, request: LogMetric, project_name: &str) -> ProjectMetricCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a logs-based metric.

Arguments

request - No description provided.

projectName - The resource name of the project in which to create the metric. Example: "projects/my-project-id". The new metric must be provided in the request.

fn sinks_delete(&self, sink_name: &str) -> ProjectSinkDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a sink.

Arguments

sinkName - The resource name of the sink to delete. Example: "projects/my-project-id/sinks/my-sink-id".

fn metrics_list(&self, project_name: &str) -> ProjectMetricListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists logs-based metrics.

Arguments

projectName - Required. The resource name of the project containing the metrics. Example: "projects/my-project-id".

fn metrics_update(&self, request: LogMetric, metric_name: &str) -> ProjectMetricUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates or updates a logs-based metric.

Arguments

request - No description provided.

metricName - The resource name of the metric to update. Example: "projects/my-project-id/metrics/my-metric-id". The updated metric must be provided in the request and have the same identifier that is specified in metricName. If the metric does not exist, it is created.