Struct google_logging2_beta1::HttpRequest

pub struct HttpRequest {
    pub status: Option<i32>,
    pub cache_validated_with_origin_server: Option<bool>,
    pub request_method: Option<String>,
    pub cache_fill_bytes: Option<String>,
    pub request_url: Option<String>,
    pub cache_lookup: Option<bool>,
    pub cache_hit: Option<bool>,
    pub referer: Option<String>,
    pub remote_ip: Option<String>,
    pub user_agent: Option<String>,
    pub request_size: Option<String>,
    pub response_size: Option<String>,
}

A common proto for logging HTTP requests.

This type is not used in any activity, and only used as part of another schema.

Fields
status: Option<i32>

The response code indicating the status of response. Examples: 200, 404.

cache_validated_with_origin_server: Option<bool>

Whether or not the response was validated with the origin server before being served from cache. This field is only meaningful if cache_hit is True.

request_method: Option<String>

The request method. Examples: "GET", "HEAD", "PUT", "POST".

cache_fill_bytes: Option<String>

The number of HTTP response bytes inserted into cache. Set only when a cache fill was attempted.

request_url: Option<String>

The scheme (http, https), the host name, the path and the query portion of the URL that was requested. Example: "http://example.com/some/info?color=red".

cache_lookup: Option<bool>

Whether or not a cache lookup was attempted.

cache_hit: Option<bool>

Whether or not an entity was served from cache (with or without validation).

referer: Option<String>

The referer URL of the request, as defined in HTTP/1.1 Header Field Definitions.

remote_ip: Option<String>

The IP address (IPv4 or IPv6) of the client that issued the HTTP request. Examples: "192.168.1.1", "FE80::0202:B3FF:FE1E:8329".

user_agent: Option<String>

The user agent sent by the client. Example: "Mozilla/4.0 (compatible; MSIE 6.0; Windows 98; Q312461; .NET CLR 1.0.3705)".

request_size: Option<String>

The size of the HTTP request message in bytes, including the request headers and the request body.

response_size: Option<String>

The size of the HTTP response message sent back to the client, in bytes, including the response headers and the response body.